split giant cell.ts file into - doc.ts (DocImpl & co) - cell.ts (Cell & co) - query-result (the proxy)
